'restructured'
