--- src/app/backoffice/services/user.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from "@angular/core/testing";
-
-import { UserService } from "./user.service";
-
-describe("UserService", () => {
-  let service: UserService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(UserService);
-  });
-
-  it("should be created", () => {
-    expect(service).toBeTruthy();
-  });
-});
