+++ src/app/public/public-routing.module.ts
@@ -10,8 +10,8 @@ import { DonationComponent } from "./pages/donations/donation/donation.component
 import { HomeComponent } from "./pages/home/home.component";
 import { PublicComponent } from "./public.component";
 import { ThanksComponent } from "./pages/donations/thanks/thanks.component";
-import { LoginFormComponent } from "./pages/auth/login-form/login-form.component";
-import { RegisterFormComponent } from "./pages/auth/register-form/register-form.component";
+import { LoginFormComponent } from "../shared/auth/login-form/login-form.component";
+import { RegisterFormComponent } from "../shared/auth/register-form/register-form.component";
 
 const routes: Routes = [
   {
