+++ src/app/shared/auth/login-form/login-form.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AbstractControl, FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Store } from '@ngrx/store';
-import { loginUser } from '../../../../state/actions/auth.actions';
+import { loginUser } from '../../../state/actions/auth.actions';
 import { checkPattern } from '../custom.validators';
 
 @Component({
