+++ src/app/shared/auth/register-form/register-form.component.spec.ts
