+++ src/app/shared/models/userState.interface.ts
@@ -1,4 +1,4 @@
-import { User } from "../backoffice/models/user";
+import { User } from "../../backoffice/models/user";
 
 export interface UserState {
   isLoged: boolean;
