+++ src/app/state/reducers/auth.reducers.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from "@ngrx/store";
 import { User } from "src/app/backoffice/models/user";
-import { UserState } from "src/app/models/userState.interface";
+import { UserState } from "src/app/shared/models/userState.interface";
 import * as actions from "../actions/auth.actions";
 
 export const initialState: UserState = { isLoged: false, user: new User };
